+++ uefi/CHANGELOG.md
@@ -20,6 +20,7 @@
 - `MemoryMap::as_raw` which provides raw access to the memory map. This is for
   example useful if you create your own Multiboot2 bootloader that embeds the
   EFI mmap in a Multiboot2 boot information structure.
+- `Mode` is now `Copy` and `Clone`
 
 ## Changed
 - `SystemTable::exit_boot_services` is now `unsafe`. See that method's
@@ -35,6 +36,8 @@
     returned type is automatically freed on the UEFI heap, as long as boot
     services are not excited. By removing the need for that explicit buffer and
     the lifetime, the API is simpler.
+- `GraphicsOutput::query_mode` is now private. Use `GraphicsOutput::modes`
+  instead.
 
 ## Removed
 - Removed the `panic-on-logger-errors` feature of the `uefi` crate. Logger